apps/gateway/requirements.txt
# SUVI Gateway Dependencies
# FastAPI and async
fastapi==0.110.0
uvicorn[standard]==0.27.1
websockets==12.0
pydantic==2.6.4
python-multipart==0.0.9

# Google Cloud SDKs
google-cloud-aiplatform==1.43.0
google-cloud-firestore==2.14.0
google-cloud-pubsub==2.19.0
google-cloud-secret-manager==2.18.0
google-cloud-bigquery==3.15.0
google-auth==2.28.0
google-auth-oauthlib==1.2.0

# A2A Protocol
aiohttp==3.9.3

# Redis for rate limiting (optional)
redis==5.0.1

# Logging
python-json-logger==2.0.7

# CORS
starlette==0.35.1
